Move Texture creation to the TileGrid
This way each Tile can be responsible for creating its own Textures.
