+++ layers.rs
@@ -7,8 +7,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use texturegl::Texture;
-use tiling::TileGrid;
+use tiling::{Tile, TileGrid};
 
 use geom::matrix::{Matrix4, identity};
 use geom::size::Size2D;
@@ -18,14 +17,8 @@ use platform::surface::{NativeCompositingGraphicsContext, NativePaintingGraphics
 use std::cell::{RefCell, RefMut};
 use std::rc::Rc;
 
-pub enum Format {
-    ARGB32Format,
-    RGB24Format
-}
-
 pub struct Layer<T> {
     pub children: RefCell<Vec<Rc<Layer<T>>>>,
-    pub tiles: RefCell<Vec<Rc<TextureLayer>>>,
     pub transform: RefCell<Matrix4<f32>>,
     pub bounds: RefCell<Rect<f32>>,
     pub tile_size: uint,
@@ -37,7 +30,6 @@ impl<T> Layer<T> {
     pub fn new(bounds: Rect<f32>, tile_size: uint, data: T) -> Layer<T> {
         Layer {
             children: RefCell::new(vec!()),
-            tiles: RefCell::new(vec!()),
             transform: RefCell::new(identity()),
             bounds: RefCell::new(bounds),
             tile_size: tile_size,
@@ -63,10 +55,6 @@ impl<T> Layer<T> {
         self.bounds.borrow_mut().size = new_size;
     }
 
-    pub fn do_for_all_buffers(&self, f: |&Box<LayerBuffer>|) {
-        self.tile_grid.borrow().do_for_all_buffers(f);
-    }
-
     pub fn add_buffer(&self, tile: Box<LayerBuffer>) {
         self.tile_grid.borrow_mut().add_buffer(tile);
     }
@@ -88,54 +76,11 @@ impl<T> Layer<T> {
     }
 
     pub fn create_textures(&self, graphics_context: &NativeCompositingGraphicsContext) {
-        // Clear all old textures.
-        self.tiles.borrow_mut().clear();
-
-        self.do_for_all_buffers(|buffer: &Box<LayerBuffer>| {
-            debug!("osmain: compositing buffer rect {}", buffer.rect);
-
-            let size = Size2D(buffer.screen_pos.size.width as int,
-                              buffer.screen_pos.size.height as int);
-
-            debug!("osmain: adding new texture layer");
-
-            // Make a new texture and bind the layer buffer's surface to it.
-            let texture = Texture::new_with_buffer(buffer);
-            debug!("COMPOSITOR binding to native surface {:d}",
-                   buffer.native_surface.get_id() as int);
-            buffer.native_surface.bind_to_texture(graphics_context, &texture, size);
-
-            // Set the layer's transform.
-            let rect = buffer.rect;
-            let transform = identity().translate(rect.origin.x, rect.origin.y, 0.0);
-            let transform = transform.scale(rect.size.width, rect.size.height, 1.0);
-
-            // Make a texture layer and add it.
-            let texture_layer = Rc::new(TextureLayer::new(texture,
-                                                          buffer.screen_pos.size,
-                                                          transform));
-            self.tiles.borrow_mut().push(texture_layer);
-        });
+        self.tile_grid.borrow_mut().create_textures(graphics_context);
     }
-}
 
-pub struct TextureLayer {
-    /// A handle to the GPU texture.
-    pub texture: Texture,
-
-    /// The size of the texture in pixels.
-    size: Size2D<uint>,
-
-    pub transform: Matrix4<f32>,
-}
-
-impl TextureLayer {
-    pub fn new(texture: Texture, size: Size2D<uint>, transform: Matrix4<f32>) -> TextureLayer {
-        TextureLayer {
-            texture: texture,
-            size: size,
-            transform: transform,
-        }
+    pub fn do_for_all_tiles(&self, f: |&Tile|) {
+        self.tile_grid.borrow().do_for_all_tiles(f);
     }
 }
